+++ app/utils/timezone.ts
@@ -1,0 +1,12 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+import {getTimeZone} from 'react-native-localize';
+
+export const isTimezoneEnabled = (config: Partial<ClientConfig>) => {
+    return config?.ExperimentalTimezone === 'true';
+};
+
+export function getDeviceTimezone() {
+    return getTimeZone();
+}
